backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.1.0
billiard==4.2.0
black==24.8.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.1
cron-descriptor==1.4.3
cryptography==43.0.0
dj-database-url==2.2.0
Django==4.2.1
django-celery-beat==2.6.0
django-cors-headers==4.1.0
django-debug-toolbar==3.4.0
django-filter==24.3
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.3
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
gcloud==0.18.3
googleapis-common-protos==1.65.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.8
inflection==0.5.1
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
kombu==5.3.7
mccabe==0.7.0
mypy-extensions==1.0.0
oauth2client==4.1.3
packaging==24.1
pathspec==0.12.1
pep8-naming==0.13.2
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.28.0
#psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.9.1
pycparser==2.22
pycryptodome==3.20.0
pyflakes==2.5.0
PyJWT==2.9.0
pyparsing==3.1.4
Pyrebase4==4.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jwt==4.1.0
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
requests==2.32.3
requests-toolbelt==0.10.1
rest-framework-simplejwt==0.0.2
rpds-py==0.19.1
rsa==4.9
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
stripe==11.1.1
